Added IInputManager interface
Slight adjustments to method signatures.  Removed some unnecessary
methods, like RemoveBindings, which was just sugar for RemoveBinding.
Updated names to reflect interface, still not happy with the Is/Are
problem for IsModifiersActive.  Removing the word "Exact" was good -
look at the doc if you're not sure what the check does (basically, the
method name shouldn't tell us how it checks, that's left to the comments
and the body of the method.)