Reworked IInputManager interface
InjectableInputManager's methods were updated to reflect the new
interface.

IInputManager:
These functions now take a PlayerIndex as a parameter:
AddBinding
RemoveBinding
ContainsBinding
ClearBinding

New functions:
RemoveBinding (overload) - now remove a binding by index (if there the
binding is overloaded) or by exact keystroke, if such is known.
ClearBinding - because a binding can now be overloaded, this method
allows you to clear all IBindings associated with a string (for a
particular player)

GetCurrentBindings - returns the list of IBindings associated with a
string.  This was added because there is no longer a 1:1 mapping between
strings and IBindings.  A string, PlayerIndex may have any number of
IBindings associated with it.

BindingsUsing - a common feature when changing keybinds in a menu is to
list currently bound keys that will be unbound if a change takes place.
This method returns a list of all bindings that the current player is
using that use a specified key.

Functions removed:
Any/All Active/Pressed/Released

Because a binding can be overloaded, there is no need to check any/all.
Instead, all bindings associated with a particular string are checked
every time you check that string.  These are no longer needed.