Removed default parameters
Default parameters suggest that a particular usage (in this case, single
player controls) is so common that it should be the assumed usage.  This
is certainly not true of something so general as input.
Further, removing defaults adds code clarity.  There were two errors in
InjectableInputManager as a result of the defaults - the
ContainsBinding(key) function was assuming a default player one, when it
should have been checking the binding for the player that it was
injected for.