Added BufferedText using IKeyboardSubscribe
By subscribing to the KeyboardDipatcher, we can push keys and commands
as they come.

I'm not sure how the RecieveSpecialInput(Keys key) method works yet, so
I'll test it out and depending on results, either add some
implementation or leave as-is.

Still need to come up with a way to ensure people initialize the
InputManager before using it (which in turn initializes the
KeyboardDispatcher, which initializes the EventInput)

Maybe constructing a InputManager before initialization will throw an
exception.