Added Chunk class
Makes it a bit easier to keep track of the global/local thing.

I need to reconsider how many "Get Global/Local Tile/Chunk" methods
there are, and possibly streamline the Manager so that it's not
unreasonable to navigate.

However, the control scheme will prefer relative chunk positions, and
the networking will likely need to deal in absolute coordinates, so...