Removed keyboard polling from DefaultInputManager
DefaultInputManager now uses the Key listener to store pressed keys in a
HashSet.  These sets (Previous/Current) are publicly readable, and
current is cleared on Update().  For this reason, it is now recommended
that you update at the END of your logic, because after input.Update(),
CurrentKeys will always be empty.

Because of the change, DefaultInputManager no longer polls the keyboard,
and instead uses these Sets to check if a key has been pressed.

This change impacted IBinding, which once again reinforces the idea that
the dependency on DefaultInputManager is too strict, and they need to be
decoupled.