Started adding XML comments
Decided to try and give this thing some decent documentation.  Turns out
there's no really easy way to have subclasses and interface
implementations inherit comments, so there's a lot of repetition.  This
wil making updating things kind of a pain, but hopefully it will reduce
the amount of comment work that needs to be done in the future.