Large rewrite of Networking
I've nearly completely moved from the original tutorial, and I've got a
decent interface for IServer.  I need to run a BasicServer and see what
we've got for exceptions when reading and losing connections, to replace
all of the "Unknown" reasons for error reporting.

It would also be a good idea to define a spec somewhere (eventually) for
the ServerEventArgs.Parameters string formatting.