M:Networking A:Log R:Old Networking
Modified:
Networking saw a bit more work, and cleaned up some of the extensions
(SafeRead was dumb, I know)

Added:
Logging.  Pretty simple class, but it should be sufficient for what I'm
hoping to do in BasicServer

Removed:
Old logging classes (BlahServer and DoCommunicate)
These needed to go, and after testing the BasicServer, it seems to
handle nicely.

Misc: Added format extension for strings, because I was sick of
realizing I was formatting a string halfway through, going back to the
beginning of the string, and adding "String.Format(" and then seeking
back to the end and adding a close parens.  Just nonsense.  Sadly, we
can't add a static Format, so I couldn't keep the capitalization
convention.  Ah well.