More robust msg passing fmt
Using messages with a fixed-size header which describes the size of the
rest of the message.  I still haven't fixed the shared resource issue
with NetworkStream, but I'll be able to tackle that now without worrying
about dropping info.

Going to do a full rewrite of the client - that thing's a mess.