Added Client class, concurrent read/write TcpClient
Wraps the read and write calls that were being used in BasicServer.  An
example server (ChatServer) had an issue when the server's read thread
was started, and then the ChatServer called GetNick (which used both
read and write) on a different thread.

There's going to be some work to do in terms of restructuring the
current read/write threads in BasicServer.