Swapped Client for TcpClient
Moved everything over to use Clients instead of TcpClients.

Swapped out dictionaries for ConcurrentDictionaries.  Because most of
our remove calls don't care whether they are the ones that removed the
object or not, and they also don't want the value that was removed, I
added an extension that lets you skip the "out" arg nonsense.