Added ServerInfoPacket
This commit should serve as a solid reminder that I need to stop rolling
my own serialization, and just use something out there.  Marc of SO fame
has a brilliant library to do this, and I'm not sure why I'm not using
it because it's genius.  When I have time, I'm gutting this whole
monstrosity and pulling that code in so that I don't have to muck around
with this nonsense.