Cleaned up Client, BasicServer
Client now takes a function that decodes packets, so that we don't have
to subclass it every time we want to change how it decodes.  This is
also good because people can now more easily extend Packet and not have
to change base code.

BasicServer was modified to compile with the Client changes, but the
stubbed PacketBuildFunc is obviously useless.