Byte array and serialization changes
- Added a bunch of extension methods for converting primatives to byte
arrays
- More Read/Write methods for streams so that we don't have to specify
startIndex, length when writing entire byte arrays
- Added WithTermChar method for strings so implementing code clearly
explains why we are appending a character to the string
- Moved GetByte methods from StringAndRegexExtensions to
ByteArrayExtensions and renamed to AsByteArray to match convention
- New interface IByteSerializeable for de/serializing objects from/to
byte arrays.  The end index is returned from the "FromByteArray" method
so that the ByteArrayReader can easily read objects.  See next commit.
- WithSizePrefix extension for byte array so that we don't have to keep
checking sizes on arrays wherever we use them.