Packet overhaul
- ChatPacket, ServerInfoPacket gone.  Either too general and boilerplate
for some cases, or too restrictive for some servers.  Either way, not
helping anyone.
- Packet now implements IByteSerializeable, so that we can easily read
using a ByteArrayReader.
- Packet now has two static functions: GetTypeFunction and
GetNameFunction.  GetTypeFunction takes an int (the 4-byte field at the
beginning of a packet byte array) and returns the string of the class
that a mapping scheme knows that packet by.  GetNameFunction does the
reverse (take a packet class name and return it's int code.)  These are
implemented so that the readonly property Packet.Type can be implemented
once and each implementation can define its own mapping between integer
codes and classes.
- The old Type property has been accordingly gutted.
- Removed unnecessary helper functions (fromInt etc)
- GetHashCode uses the ByteArray's hash code.
- Equals uses GetHashCode