IByteSerializeable, ByteArray Builder/Reader
IByteSerializeable: Serialize to a byte array, and load from one.
FromByteArray returns the index of the end of the object deserialized,
so that ByteArrayReader can easily track objects of arbitrary length.
Should return a negative integer if the data starting at startIndex is
not an appropriate type object (trying to deserialize a car object, but
data is actually an animal object)

ByteArrayBuilder: StringBuilder for byte arrays.  Supports
IByteSerializeable interface

ByteArrayReader: Allows you to sequentially read objects from a byte
array.  Each read starts from the end of the previous, so that you don't
have to manually track the pointer.  In some cases, this removes the
need to know the length of the object entirely (e.g. well-formatted
strings)