Client can push read events
BasicServer and ChatClient/ChatServer were using threads all over so
that they could read and write to a Client.  This is silly because
Client is intended to remove the need for that stuff.  Now, Client has
an event OnReadPacket which will alert any subscribers when a new packet
is available for reading.  It does not provide the packet itself, but
lets those subscribers know that the packet is available for reading.