BasicServer reworked to use Client Read push
Instead of starting a new thread that polls every client for each
client, the server now subscribes to each client's OnReadPacket event.

This is a significant performance improvement for cases where the client
count is high, or clients are often disconnecting.

To prevent disconnected clients from streaming garbage packets to the
server, the server will check that the client is still in the
ClientTable - if not, it unsubscribes itself from the bad client's
OnReadPacket event.